Minor tweaks and thoroughly checked nm-folder Minor tweaks have to do with variable order, which happened to be important. Thoroughly checke nm-folder.js functions and next work is on nm-file.js.
Think of a solution to get the variables of path just right, since these
have to be re-initialized whenever dreamweaver insert an object. This is
due to all dreamweaver tabs sharing this script file.

Solutions to this problem could be immediate invocation/closure or just
an var_init function where the values will be initiated globally, this
could also prove to be a problem when we take advantage of the
persistend of objects and variables...